% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{alleeEffectDynamics}
\alias{alleeEffectDynamics}
\title{Allee effect model}
\usage{
alleeEffectDynamics(launchBrowser = FALSE)
}
\arguments{
\item{launch.browser}{If this is FALSE (the default option), the shiny app will be launched using default settings (usually a new window created by RStudio). If set to TRUE, the app will be opened in the system's default web browser.}
}
\description{
Launches a shiny app that allows the user to change the parameters of an allee effect model
of population growth and observe the resulting dynamics.
}
\details{
The ordinary differential equation for this model is \eqn{dN/dt = rN(1-N/K)(N/T-1)},
where \eqn{N} is the population size, \eqn{K} the carrying capacity,
 and \eqn{T} the threshold population size below which the population goes extinct.
}
