% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{epidemiologyDynamics}
\alias{epidemiologyDynamics}
\title{Epidemiological ODE models}
\usage{
epidemiologyDynamics(launchBrowser = FALSE)
}
\arguments{
\item{launch.browser}{If this is FALSE (the default option), the shiny app will be launched using default settings (usually a new window created by RStudio). If set to TRUE, the app will be opened in the system's default web browser.}
}
\description{
Launches a shiny app that interatively plots the dynamics of three commonly used epidemiological models: the SI, SIS and SIR model.
}
\details{
The differential equation systems for the three models are as follows:

SI model: \eqn{dS/dt = \nu - \betaIS - \muS, dI/dt = \betaIS - (\mu + \alpha)I}

SIS model: \eqn{dS/dt = \nu - \betaIS - \muS + \gammaI, dI/dt = \betaIS - (\mu + \alpha - \gamma)I}

SIR model: \eqn{dS/dt = \betaIS, dI/dt = \betaIS - \gammaI, dR/dt = \gammaI}
}
