% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{haploidSelection}
\alias{haploidSelection}
\title{Haploid selection model}
\usage{
haploidSelection(launchBrowser = FALSE)
}
\arguments{
\item{launch.browser}{If this is FALSE (the default option), the shiny app will be launched using default settings (usually a new window created by RStudio). If set to TRUE, the app will be opened in the system's default web browser.}
}
\description{
Launches a shiny app that allows the user to change the parameters of a model of natural selection in a haploid populatiom, and observe the resulting dynamics.
}
\details{
The recurrence equation of this model is \eqn{p'=(1+s)p/(1+sp)}, where p is the frequency of the focal allele (with 1-p the frequency of the other allele), and s is the selection coefficient.
}
