% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{poissonProcess}
\alias{poissonProcess}
\title{Poisson process}
\usage{
poissonProcess(launchBrowser = FALSE)
}
\arguments{
\item{launch.browser}{If this is FALSE (the default option), the shiny app will be launched using default settings (usually a new window created by RStudio). If set to TRUE, the app will be opened in the system's default web browser.}
}
\description{
Launches a shiny app illustrating the Poisson process for different rates.
}
\details{
The Poisson process takes place over a fixed time period of 100 units.
}
